Complete testing of LogEvent. Experiment with Tail and Scanner
Complete parsing of log events from log files. At the moment once a
ERROR log event is found a Caused By is searched for automatically
and a loose association is created with the log event and caused by.

Scanner will be used in future to preload the database with log events
Tail will be used to watch a file