Add Stat calc and notify based on stats
Add table[error_day_summary] for summaries of error events
Add table[error_stats] for mean, std dev, total and count stats based on
error events
After items are loaded from a logs directory these tables are filled
with current stat data for use in future events
After all stats have been calculated a log file is watched. Every item
that appears in this watched file is sent to a event processing go
routine which compares the event with historic stats on the particular
item. If the error count for the day exceeds the std dev + mean the
decision is made to notify someone

Currently notifying is not implemented