Logs are parsed concurrently. Fix errord parsing
ErrorD parsing of Caused By reported false postives with regard to add
extract Caused By Exception and Detail to the Event. Which caused the
parser to not add any events to the store. Added test cases for the
above scenario with AddCausedByLine.

ParseCausedBy not returns the extracted parts and an error. This makes
its purpose more clear as previously it added the extracted parts to the
event, which made the AddCausedBy method weird since it didn't do it.
The AddCausedBy line now adds the extracted parts to the event